### 2018-02-20 #591 @whedon
[PRE REVIEW]: A reusable tree-based web-visualization to browse EDAM ontology, and contribute to it.
**Submitting author:** @bryan-brancotte (<a href="http://orcid.org/0000-0001-8669-5525">Bryan Brancotte</a>)
**Repository:** <a href="https://github.com/IFB-ElixirFr/edam-browser" target ="_blank">https://github.com/IFB-ElixirFr/edam-browser</a>
**Version:** v1.0.0
