### 2018-02-20 090ba18
replacing checkmark and cross by their html version
