### 2018-02-20 e209531
merging edition figures
