### 2018-02-20 db68675
removing checkmark as not working with joss pdf generation dirty attempt to reduce image height in joss pdf generation
