### 2018-02-20 d8169b3
working on figure size
