### 2018-02-20 f1be360
attempt to bring the json schema in one piece in pdf
