### 2018-02-20 #591 @bryan-brancotte
Hi
The labels should be updated, Python should be removed and JavaScript added.
Best,
