### 2018-02-20 024cbf1
attempt to bring the json schema in one piece in pdf
