### 2018-02-27 7dd8b9d
minor typo corrections
