### 2018-02-28 #591 @bryan-brancotte
Sorry, for the inconvenience @arfon, I though that I had to do an action to initiate the search of editor and reviewer.

Regarding the labels, Python should be removed and JavaScript added.

Cheers
Bryan
