### 2018-03-07 #591 @arfon
@bryan-brancotte - many thanks for your submission to JOSS. This looks like an interesting project.

The paper as it stands, is substantially longer than we usually allow for JOSS submissions. In particular, much of this content looks like it might be more appropriate for for the software documentation.

Could you please take a look at the ['What should my paper contain'](http://joss.theoj.org/about#paper_structure) section of the author guidelines and revise the paper?
