### 2018-03-12 b6b32d7
Adding paper for short version
