+++ paper.md
@@ -1,0 +1,203 @@
+---
+title: 'A reusable tree-based web-visualization to browse EDAM ontology, and contribute to it.'
+tags:
+- javascript
+- bioinformatics
+- d3js
+- d3.js
+- ontology
+- visualization
+- edamontology
+- ontology
+authors:
+- name: Bryan Brancotte
+  orcid: 0000-0001-8669-5525
+  affiliation: "1" # (Multiple affiliations must be quoted)
+- name: Christophe Blanchet
+  affiliation: 2
+- name: Hervé Ménager
+  orcid: 0000-0002-7552-1009
+  affiliation: 1
+affiliations:
+- name: Centre d'Informatique pour la Biologie, C3BI, Institut Pasteur, Paris, France
+  index: 1
+- name: French Institute of Bioinformatics, CNRS IFB-Core, Gif-sur-Yvette, France
+  index: 2
+date: 01 February 2018
+bibliography: paper_resources/paper.bib
+---
+
+# Summary
+
+## Abstract
+
+Labelling, indexing and describing a Bioinformatics resource, whether it is a software, a database, or a service is of a great help when it comes to promoting it to various user communities. As an example, the ELIXIR bio.tools [@ison2015tools] registry contains more than ten thousands software and service entries. In this context, the use of controlled vocabularies to describe the resources is of a paramount importance. In bio.tools, this need is addressed by the EDAM Ontology [@ison2013edam], which proposes a controlled vocabulary hierarchically organized around four axes which describe types of data, formats, operations and topics.
+
+We here present the EDAM Browser, a client-side web-based visualization javascript widget that provides an interface to navigate EDAM. This browser is tailored to the needs of EDAM users that might not be ontology experts. It can, among other things, be used to help describing resources, and to facilitate and foster community contributions to EDAM. Throughout the rest of this paper, (i) we review related work and features that need to be addressed; (ii) we describe how these features are addressed by the EDAM Browser; finally (iii) we discuss its reusability.
+
+## Related work
+
+We collected from EDAM users and authors features that a browser should have to fit their needs. Hereafter we describe collected criteria and their presence or absence in the following platforms: (i) BioPortal[@whetzel2011bioportal], (ii) OLS - Ontology Lookup Service[@jupp2015new], (iii)  WebProt&eacute;g&eacute;[@tudorache2013webprotege] and (iv) AberOWL[@hoehndorf2015aber].
+
+| id  | Feature                                | AberOWL                    | BioPortal          | OLS                        | WebProtégé                 |
+|:---:|----------------------------------------|:--------------------------:|:------------------:|:--------------------------:|:--------------------------:|
+| 1   | EDAM is present                        | yes                        | yes                | yes                        | yes                        |
+| 2   | Publicly available                     | yes                        | yes                | yes                        | registration               |
+| 3   | Display of multiple parents            | no                         | no                 | yes                        | no                         |
+| 4   | Can be integrated in external websites | no                         | yes                | yes                        | no                         |
+| 5   | Facilitate community contribution      | no                         | limited            | no                         | yes                        |
+| 6   | Use edited version of EDAM             | no                         | no                 | no                         | yes                        |
+| 7   | Local installation                     | yes      (w/ admin rights) | yes      (as a VM) | yes      (w/ admin rights) | yes      (w/ admin rights) |
+| 8   | Links to annotated resources           | no                         | no                 | no                         | no                         |
+
+
+##### Criteria 1 and 2 - "EDAM is present and publicly available"
+
[…]
