+++ paper.md
@@ -35,161 +35,61 @@ Labelling, indexing and describing a Bioinformatics resource, whether it is a so
 
 We here present the EDAM Browser, a client-side web-based visualization javascript widget that provides an interface to navigate EDAM. This browser is tailored to the needs of EDAM users that might not be ontology experts. It can, among other things, be used to help describing resources, and to facilitate and foster community contributions to EDAM. Throughout the rest of this paper, (i) we review related work and features that need to be addressed; (ii) we describe how these features are addressed by the EDAM Browser; finally (iii) we discuss its reusability.
 
-## Related work
-
-We collected from EDAM users and authors features that a browser should have to fit their needs. Hereafter we describe collected criteria and their presence or absence in the following platforms: (i) BioPortal[@whetzel2011bioportal], (ii) OLS - Ontology Lookup Service[@jupp2015new], (iii)  WebProt&eacute;g&eacute;[@tudorache2013webprotege] and (iv) AberOWL[@hoehndorf2015aber].
-
-| id  | Feature                                | AberOWL                    | BioPortal          | OLS                        | WebProtégé                 |
-|:---:|----------------------------------------|:--------------------------:|:------------------:|:--------------------------:|:--------------------------:|
-| 1   | EDAM is present                        | yes                        | yes                | yes                        | yes                        |
-| 2   | Publicly available                     | yes                        | yes                | yes                        | registration               |
-| 3   | Display of multiple parents            | no                         | no                 | yes                        | no                         |
-| 4   | Can be integrated in external websites | no                         | yes                | yes                        | no                         |
-| 5   | Facilitate community contribution      | no                         | limited            | no                         | yes                        |
-| 6   | Use edited version of EDAM             | no                         | no                 | no                         | yes                        |
-| 7   | Local installation                     | yes      (w/ admin rights) | yes      (as a VM) | yes      (w/ admin rights) | yes      (w/ admin rights) |
-| 8   | Links to annotated resources           | no                         | no                 | no                         | no                         |
-
-
-##### Criteria 1 and 2 - "EDAM is present and publicly available"
-
-The first two criteria designate wether these interfaces allow to navigate EDAM, and to do so in a anonymous manner. The access to the web interface should be as simple and fast as possible, especially for users who do not need ellaborate functionalities. On the one hand the AberOWL, BioPortal and OLS platforms integrate EDAM and let unauthenticated users browse it. On the other hand, WebProt&eacute;g&eacute; is meant for ontology edition. It asks users to register before uploading its ontology, or browsing the [latest stable version](https://webprotege.stanford.edu/#projects/98640503-a37d-4404-84da-caf30fadd685/edit/Classes) of EDAM.
-
-##### Criteria 3 - "Display of multiple parents"
-
[…]
