+++ paper_resources/paper.bib
@@ -64,7 +64,6 @@
   year={2015},
   publisher={BioMed Central}
 }
-
 @misc{brancotte2017biosphere,
   TITLE = {{Biosphere Web : un portail haut niveau pour une utilisation bioinformatique des clouds}},
   AUTHOR = {Brancotte, Bryan and Bedri, Mohamed and Lorenzo, Jonathan and Perrin Sandrine and Séné, Frédéric and Sepou Ngaïlo, Awa and Blanchet, Christophe and Gibrat, Jean-François},
@@ -73,4 +72,10 @@
   HOWPUBLISHED = {{JOBIM (Journ{\'e}es Ouvertes Biologie Informatique Math{\'e}matiques)}},
   YEAR = {2017},
   MONTH = Jul,
+}
+@inproceedings{xiang2011ontobee,
+  title={Ontobee: A Linked Data Server and Browser for Ontology Terms.},
+  author={Xiang, Zuoshuang and Mungall, Chris and Ruttenberg, Alan and He, Yongqun},
+  booktitle={ICBO},
+  year={2011}
 }
