+++ paper_resources/paper.bib
@@ -66,7 +66,7 @@
 }
 @misc{brancotte2017biosphere,
   TITLE = {{Biosphere Web : un portail haut niveau pour une utilisation bioinformatique des clouds}},
-  AUTHOR = {Brancotte, Bryan and Bedri, Mohamed and Lorenzo, Jonathan and Perrin Sandrine and Séné, Frédéric and Sepou Ngaïlo, Awa and Blanchet, Christophe and Gibrat, Jean-François},
+  AUTHOR = {Brancotte, Bryan and Bedri, Mohamed and Lorenzo, Jonathan and Perrin Sandrine and S{\'e}n{\'e}, Fr{\'e}d{\'e}ric and Sepou Nga{\"i}lo, Awa and Blanchet, Christophe and Gibrat, Jean-Fran\c{c}ois},
   URL = {https://www.youtube.com/watch?v=gaopunfIqnc},
   NOTE = {Demo},
   HOWPUBLISHED = {{JOBIM (Journ{\'e}es Ouvertes Biologie Informatique Math{\'e}matiques)}},
