### 2018-03-13 ff01b1f
additional work on the paper
reduce number of sections and summarize
