+++ paper.md
@@ -33,71 +33,37 @@ bibliography: paper_resources/paper.bib
 
 Labelling, indexing and describing a Bioinformatics resource, whether it is a software, a database, or a service is of a great help when it comes to promoting it to various user communities. As an example, the ELIXIR bio.tools [@ison2015tools] registry contains more than ten thousands software and service entries. In this context, the use of controlled vocabularies to describe the resources is of a paramount importance. In bio.tools, this need is addressed by the EDAM Ontology [@ison2013edam], which proposes a controlled vocabulary hierarchically organized around four axes which describe types of data, formats, operations and topics.
 
-We here present the EDAM Browser, a client-side web-based visualization javascript widget that provides an interface to navigate EDAM. This browser is tailored to the needs of EDAM users that might not be ontology experts. It can, among other things, be used to help describing resources, and to facilitate and foster community contributions to EDAM. Throughout the rest of this paper, (i) we review related work and features that need to be addressed; (ii) we describe how these features are addressed by the EDAM Browser; finally (iii) we discuss its reusability.
+We here present the EDAM Browser, a client-side web-based visualization javascript widget that provides an interface to navigate EDAM. This browser is tailored to the needs of EDAM users that might not be ontology experts. It can, among other things, be used to help describing resources, and to facilitate and foster community contributions to EDAM. The EDAM Browser allows users to explore it with an interface tailored to its structure and properties. Its interface is not designed to be a generic ontology navigation and edition platform, a goal already achieved by many other systems (AberOWL[@hoehndorf2015aber], BioPortal[@whetzel2011bioportal], OLS - Ontology Lookup Service[@jupp2015new], Ontobee[@xiang2011ontobee], WebProt&eacute;g&eacute;[@tudorache2013webprotege]).
 
-## Main functionalities
+Rather, it aims at provide features requested by most users and contributors, who might not be ontology experts or maintainers, which we detail below.
 
-We collected from EDAM users and authors features that a browser should have to fit their needs. We evaluated different platforms, and found that none of them propose all the features asked by users. Hereafter are the evaluated platforms: 
-* AberOWL[@hoehndorf2015aber]
-* BioPortal[@whetzel2011bioportal]
-* OLS - Ontology Lookup Service[@jupp2015new]
-* Ontobee[@xiang2011ontobee]
-* WebProt&eacute;g&eacute;[@tudorache2013webprotege].
+## Availability
 
-Hereafter we list features and detail how we address them. 
+The EDAM browser is available publicly and anonymously at https://ifb-elixirfr.github.io/edam-browser/. In addition to this, its lightweight architecture makes it easy to download and run on any server or personal computer, either as a local HTML file or on a web server. It is possible to integrate the EDAM Browser and its tree representation in external websites and applications, providing a simple way for third party websites to promote EDAM-labeled resources. Both the autocomplete input field and the tree visualization are re-usable: a demonstration code is available [here](https://ifb-elixirfr.github.io/edam-browser/demo.html), showing how the tree can be integrated, how the user can interact with the tree, and how to programmaticaly interact with the tree in JavaScript.
 
-##### EDAM is present and publicly available
+## Information display
 
-The EDAM Browser is publicly available and does not require authentication. 
+As much as possible, the user interface aims at simplicity and relevance to the specific domain of EDAM. The creation of an interface that displays all of the information necessary to users, and avoids the use of ontology development jargon is a major goal of this project. 
 
-##### Display of multiple parents
-
[…]
