### 2018-03-13 #591 @bryan-brancotte
Hi @arfon
We re submitted a condensed version of the paper (-60%) which is now compliant with 'What should my paper contain'. Note that the compilation never ends.
Best regards
