### 2018-03-13 #591 @arfon
Here's the compiled PDF: [10.21105.joss.00591.pdf](https://github.com/openjournals/joss-reviews/files/1807659/10.21105.joss.00591.pdf)
