### 2018-03-14 4a4439f
Update paper.md
