### 2018-03-14 #591 @bryan-brancotte
Hi @arfon
Thanks for the pdf, I have done minor updates in the paper since its production, but the result it ok for us.
