### 2018-03-25 #591 @pjotrp
@alexgarciac offers to review the Edam browser.
