### 2018-03-25 #591 @arfon
> Nice try. @arfon, two editors not allowed. @whedon is silent. @jasonclark I can help find a reviewer, or take over as an editor if you prefer.

Yes, this is expected. https://github.com/openjournals/whedon-api/pull/25 only added support for multiple _reviewers_.
