### 2018-04-23 #698 @pjotrp
Dear @bryan-brancotte, sorry for the delay. Mostly because someone stopped communicating - hard to deal with that. But we'll move on now. To expedite the work can you add the bibliography and compile the PDF until it works? Also please go through above check list and ascertain the reviewer can check the boxes. Much faster if you make sure we can tick them than wait for a back and forth. Meanwile @alexgarciac can you go through the JOSS tips for review at http://joss.theoj.org/about#reviewer_guidelines. Thanks!
