### 2018-04-23 #698 @hmenager
Hi @pjotrp, thanks a lot for handling this. Regarding the bibliography, the compilation error seems weird, because the bib file is actually there as far as I can tell (https://github.com/IFB-ElixirFr/edam-browser/blob/master/paper_resources/paper.bib). On the previous round @arfon actually provided the pdf manually. Is there something we should correct in the markdown or the bibtex to avoid the compilation errors? Thanks again!
