### 2018-04-24 #698 @arfon
@hmenager - the issue here is that there are two papers so @whedon doesn't know what to do:

> 1. tmp/698/paper.md
> 2. tmp/698/paper_resources/long_paper.md

I've compiled this locally and uploaded here: [10.21105.joss.00698.pdf](https://github.com/openjournals/joss-reviews/files/1940528/10.21105.joss.00698.pdf)
