### 2018-04-24 #698 @arfon
> The second paper should be renamed?

Yes. The regex goes looking for anything with `paper.md` in the title. Honestly it doesn't really matter, you just can't ask Whedon to auto-compile the paper here.
