### 2018-04-25 #698 @hmenager
@pjotrp I will check that carefully with @bryan-brancotte, and get back to you real soon. From quickly checking the list above, I think we miss a release and the contribution guidelines. We'll ping you as soon as these are done.
