### 2018-04-30 324bc0d
jshint compliance
