+++ js/tree-reusable-d3.js
@@ -5,12 +5,12 @@
 function interactive_tree() {
     var debug=false,
         margin = {top: 10, right: 50, bottom: 10, left: 80},
-        voidHandler=function(name){return function(){if(debug)console.log(name);}},
-        voidHandlerStrBlank=function(name){return function(){if(debug)console.log(name);return"";}},
+        voidHandler=function(name){return function(){if(debug)console.log(name);};},
+        voidHandlerStrBlank=function(name){return function(){if(debug)console.log(name);return"";};},
         initiallySelectedElementHandler = voidHandler("initiallySelectedElementHandler"),
         addingElementHandler = voidHandler("addingElementHandler"),
         openElementHandler = voidHandler("openElementHandler"),
-        clickedElementHandler = function(d){if(debug)console.log(identifierAccessor(d))},
+        clickedElementHandler = function(d){if(debug)console.log(identifierAccessor(d));},
         removingElementHandler = voidHandler("removingElementHandler"),
         loadingDoneHandler = voidHandler("loadingDoneHandler"),
         metaInformationHandler = voidHandler("metaInformationHandler"),
@@ -33,7 +33,7 @@ function interactive_tree() {
                     "</div>"+
                     "<div class=\"panel-body\">"+
                     "Identifier: "+identifierAccessor(d)+
-                    "</div></div>"
+                    "</div></div>";
         },
         preTreatmentOfLoadedTree=function(tree){return  tree;},
         tooltipEnabled=false,
@@ -51,7 +51,7 @@ function interactive_tree() {
         duration = 500,
         update,
         reset,
-        id=0
+        id=0,
         identifierToElement={};
 
     function chart(selection){
@@ -77,7 +77,7 @@ function interactive_tree() {
                     d3.event.transform.y = Math.max(dimG.height*-0.8/2,d3.event.transform.y);
                     d3.event.transform.x = Math.min((dimS.width-50)*0.8,d3.event.transform.x);
                     d3.event.transform.y = Math.min((dimS.height+dimG.height/2)*0.8,d3.event.transform.y);
-                    vis.attr("transform", d3.event.transform)
+                    vis.attr("transform", d3.event.transform);
                 });
 
             var svg = d3.select(this).append("svg:svg")
@@ -111,7 +111,7 @@ function interactive_tree() {
                     shift=50;
                 var t = d3.zoomIdentity.translate(shift,$(target_selector).height()/2).scale(1);
                 svg.call(zoom.transform, t);
-            }
+            };
 
             update = function (source) {
                 if (treeSelectedElementAncestors == null){
@@ -137,7 +137,7 @@ function interactive_tree() {
                 var nodeEnter = node.enter().append("svg:g")
                     .attr("class", "node")
                     .attr("transform", function(d) {
-                        return "translate(" + source.y0 + "," + source.x0 + ")";
+                        return "translate(" + (source.y0 || source.y) + "," + (source.x0 || source.x) + ")";
                     })
                     .on("click", function(d,i) { handleClick(d);});
 
@@ -151,8 +151,7 @@ function interactive_tree() {
                 nodeEnter.append("svg:circle")
                     .attr("r", 1e-6)
                     .attr("class", function(d) {
-                        return (d._children || d.children ? "has-children " : " ")
-                        + additionalCSSClassForNode(d);
+                        return (d._children || d.children ? "has-children " : " ") + additionalCSSClassForNode(d);
                     })
                     .on("mouseover", function(d) {
                         if (!tooltipEnabled) return;
@@ -229,7 +228,7 @@ function interactive_tree() {
                 var linkEnter = link.enter().insert("svg:path", "g")
                     .attr("data-id", function(d) { return d.__d3js_id; })
[…]
