+++ js/biosphere.api.js
@@ -8,25 +8,25 @@ function biosphere_api(){
 
     //getter for nothing
     var get_for_nothing=function (name){
-        var getter = function(){}
-        getter.count=function(callback){}
+        var getter = function(){};
+        getter.count=function(callback){};
         getter.is_enabled=function(){
             return false;
-        }
-        getter.get_url=function(){}
-        getter.get_api_url=function(){}
+        };
+        getter.get_url=function(){};
+        getter.get_api_url=function(){};
         return getter;
-    }
+    };
 
     // generic getter
     api.get_for=function (branch, name, uri, node){
         if (uri.indexOf("edam")==-1)
             return get_for_nothing();
         var term_id=uri.substring(uri.lastIndexOf("/")+1);
-        var branch = term_id.substring(0,term_id.lastIndexOf("_"));
+            branch = term_id.substring(0,term_id.lastIndexOf("_"));
         var number = term_id.substring(term_id.lastIndexOf("_")+1);
         //getter object
-        var getter = function(){}
+        var getter = function(){};
         //function to count the number of tools associated
         getter.count=function(callback){
             return $.ajax({
@@ -42,10 +42,11 @@ function biosphere_api(){
                     var tools_row=data.content.split("href=\"/catalogue/tool/");
                     var appliances=[];
                     var tools=[];
-                    for(var i=1;i<appliances_row.length;i++){
+                    var i;
+                    for(i=1;i<appliances_row.length;i++){
                         appliances.push(appliances_row[i].match(/[^>]*>([^<]*)[<]/)[1].trim());
                     }
-                    for(var i=1;i<tools_row.length;i++){
+                    for(i=1;i<tools_row.length;i++){
                         tools.push(tools_row[i].match(/[^>]*>([^<]*)[<]/)[1].trim());
                     }
                     callback(
@@ -70,36 +71,36 @@ function biosphere_api(){
                     callback(-1,[],textStatus);
                 },
             });
-        }
+        };
         //is the count function enabled
         getter.is_enabled=function(){
             return true;
-        }
+        };
         //get the url returning the tools for human
         getter.get_url=function(){
             return "https://biosphere.france-bioinformatique.fr/edamontology/"+branch+"/"+number+"/";
-        }
+        };
         //get the url returning the tools for api call
         getter.get_api_url=function(){
             return getter.get_url()+"?media=json&included=True";
-        }
+        };
         return getter;
-    }
+    };
     //getter for topics
     api.get_for_topic=function (name){
-        return api.get_for("topic",name)
-    }
+        return api.get_for("topic",name);
+    };
     //getter for operations
     api.get_for_operation=function (name){
-        return get_for_nothing()
-    }
+        return get_for_nothing();
+    };
     //getter for data
     api.get_for_data=function (name){
-        return get_for_nothing()
-    }
+        return get_for_nothing();
+    };
     //getter for format
     api.get_for_format=function (name){
-        return get_for_nothing()
-    }
+        return get_for_nothing();
+    };
     return api;
 }
