### 2018-04-30 #698 @bryan-brancotte
Hi @pjotrp, we renamed the long_paper.md and reviewed all the boxes: they should be good to check.
