+++ .travis.yml
@@ -1,3 +1,5 @@
 language: node_js
 node_js:
  - "node"
+script:
+  - npm test
