### 2018-05-02 #698 @pjotrp
@alexgarciac please start the review by ticking the boxes at the top of this page.
