### 2018-05-21 #698 @alexgarciac
Hi, I have checked the appropriate boxes. the ones not checked have comments in the issue tracker of the edam prj.
