### 2018-05-22 #698 @pjotrp
Thans @alexgarciac for your very thorough review. @bryan-brancotte I think it is a good idea to look at all suggestions and see if you can improve the paper. The goal is to present your work in the best way so people start using it! I don't think it is a lot of work, so it is close to being accepted with minor revisions only.

Of the missing check boxes, I think only the one on community guidelines to be really important. See if we can check that box. Also DOIs and installation instructions would be a good idea.
