+++ paper_resources/paper.bib
@@ -6,7 +6,8 @@
   number={D1},
   pages={D38--D47},
   year={2015},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkv1116}
 }
 
 @article{whetzel2011bioportal,
@@ -17,7 +18,8 @@
   number={suppl\_2},
   pages={W541--W545},
   year={2011},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkr469}
 }
 @inproceedings{jupp2015new,
   title={A new Ontology Lookup Service at EMBL-EBI.},
@@ -34,7 +36,8 @@
   number={10},
   pages={1325--1332},
   year={2013},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btt113}
 }
 @misc{
   d3js,
@@ -52,17 +55,26 @@
   number={1},
   pages={89--99},
   year={2013},
-  publisher={IOS Press}
+  publisher={IOS Press},
+  doi={10.3233/SW-2012-0057}
 }
 @article{hoehndorf2015aber,
-  title={Aber-OWL: a framework for ontology-based data access in biology},
-  author={Hoehndorf, Robert and Slater, Luke and Schofield, Paul N and Gkoutos, Georgios V},
-  journal={BMC bioinformatics},
-  volume={16},
-  number={1},
-  pages={26},
-  year={2015},
-  publisher={BioMed Central}
+author="Hoehndorf, Robert
+and Slater, Luke
+and Schofield, Paul N.
+and Gkoutos, Georgios V.",
+title="Aber-OWL: a framework for ontology-based data access in biology",
+journal="BMC Bioinformatics",
+year="2015",
+month="Jan",
+day="28",
+volume="16",
+number="1",
+pages="26",
+abstract="Many ontologies have been developed in biology and these ontologies increasingly contain large volumes of formalized knowledge commonly expressed in the Web Ontology Language (OWL). Computational access to the knowledge contained within these ontologies relies on the use of automated reasoning.",
+issn="1471-2105",
+doi="10.1186/s12859-015-0456-9",
+url="https://doi.org/10.1186/s12859-015-0456-9"
 }
 @misc{brancotte2017biosphere,
   TITLE = {{Biosphere Web : un portail haut niveau pour une utilisation bioinformatique des clouds}},
@@ -77,5 +89,6 @@
   title={Ontobee: A Linked Data Server and Browser for Ontology Terms.},
   author={Xiang, Zuoshuang and Mungall, Chris and Ruttenberg, Alan and He, Yongqun},
   booktitle={ICBO},
-  year={2011}
+  year={2011},
+  doi={10.1093/nar/gkw918}
 }
