### 2018-05-25 #698 @bryan-brancotte
Hi @pjotrp
Thank to @alexgarciac for all those comments. I had a quite dense week and going to be offline for two weeks so I will not be able to take into account the suggestions in the mean time. As soon as I am back online be assured that it will be on the top of my todo list.
Best regards
