### 2018-06-14 b36307d
add URL for OLS paper which doesn't have any DOI
DOIs requested by #16
