+++ paper_resources/paper.bib
@@ -26,7 +26,8 @@
   author={Jupp, Simon and Burdett, Tony and Leroy, Catherine and Parkinson, Helen E},
   booktitle={SWAT4LS},
   pages={118--119},
-  year={2015}
+  year={2015},
+  url={{http://ceur-ws.org/Vol-1546/paper_29.pdf}}
 }
 @article{ison2013edam,
   title={EDAM: an ontology of bioinformatics operations, types of data and identifiers, topics and formats},
