### 2018-06-14 f59b341
Update paper.md
adressed comments from https://github.com/IFB-ElixirFr/edam-browser/issues/16#issuecomment-390509294 in #16
