### 2018-06-14 35eceae
Update paper.md
adressing comments from https://github.com/IFB-ElixirFr/edam-browser/issues/16#issuecomment-390510684 from #16
