### 2018-06-15 d134cc8
Restructuring summary
following https://github.com/IFB-ElixirFr/edam-browser/issues/16#issuecomment-390510684, but not only
