### 2018-06-15 bb9b172
reorganizing Information display section, enriching image.
