+++ paper.md
@@ -47,7 +47,7 @@ The interface permits the navigation between different axes of the ontology, bas
 
 We also take into account the specificities of the structure of EDAM: while being represented as a tree, it is in fact a directed acyclic graph, meaning that a term can have more than one parent. In order to improve readability when a term is selected (1) all the term's positions are shown; and (2) all paths from the root node are highlighted. A good example of this display is the [Phylogeny topic](https://ifb-elixirfr.github.io/edam-browser/#topic_0084)) (cf Fig. 1). 
 
-On the right part of Figure 1, details of the selected concept are displayed. One last salient feature is the presentation of the usage of the concept in annotated resource collections, such as bio.tools, [BioSphere](https://biosphere.france-bioinformatique.fr/catalogue/) [@brancotte2017biosphere], [BioWeb](http://bioweb.pasteur.fr) and [TeSS](https://tess.elixir-europe.org).
+On the right part of Figure 1, details of the selected concept are displayed. One last salient feature is the presentation of the usage of the concept in annotated resource collections, such as [bio.tools](https://bio.tools), [BioSphere](https://biosphere.france-bioinformatique.fr/catalogue/) [@brancotte2017biosphere], [BioWeb](http://bioweb.pasteur.fr) and [TeSS](https://tess.elixir-europe.org).
 
 ## Performance and flexibility
 
