+++ paper_resources/paper.bib
@@ -105,3 +105,5 @@ doi = {10.1093/nar/gkw343},
 URL = {http://dx.doi.org/10.1093/nar/gkw343},
 eprint = {/oup/backfile/content_public/journal/nar/44/w1/10.1093_nar_gkw343/3/gkw343.pdf}
 }
+@article{ashburner2000go, title={Gene Ontology: tool for the unification of biology}, volume={25}, url={http://dx.doi.org/10.1038/75556}, DOI={10.1038/75556}, number={1}, journal={Nature Genetics}, publisher={Springer Nature}, author={Ashburner, Michael and Ball, Catherine A. and Blake, Judith A. and Botstein, David and Butler, Heather and Cherry, J. Michael and Davis, Allan P. and Dolinski, Kara and Dwight, Selina S. and Eppig, Janan T. and et al.}, year={2000}, month={May}, pages={25–29}}
+ @article{go2016go, title={Expansion of the Gene Ontology knowledgebase and resources}, volume={45}, url={http://dx.doi.org/10.1093/nar/gkw1108}, DOI={10.1093/nar/gkw1108}, number={D1}, journal={Nucleic Acids Research}, publisher={Oxford University Press (OUP)}, year={2016}, month={Nov}, pages={D331–D338}} 
