### 2018-06-15 2923428
Update paper.md
reformulate and trim paper
