### 2018-06-15 b3473cf
Update paper.md
more paper updates/trimming
