### 2018-06-15 b9f8a22
restructure and simplify the paper to comply more with JOSS guidelines (#16)
