+++ paper_resources/paper.bib
@@ -106,4 +106,16 @@ URL = {http://dx.doi.org/10.1093/nar/gkw343},
 eprint = {/oup/backfile/content_public/journal/nar/44/w1/10.1093_nar_gkw343/3/gkw343.pdf}
 }
 @article{ashburner2000go, title={Gene Ontology: tool for the unification of biology}, volume={25}, url={http://dx.doi.org/10.1038/75556}, DOI={10.1038/75556}, number={1}, journal={Nature Genetics}, publisher={Springer Nature}, author={Ashburner, Michael and Ball, Catherine A. and Blake, Judith A. and Botstein, David and Butler, Heather and Cherry, J. Michael and Davis, Allan P. and Dolinski, Kara and Dwight, Selina S. and Eppig, Janan T. and et al.}, year={2000}, month={May}, pages={25–29}}
- @article{go2016go, title={Expansion of the Gene Ontology knowledgebase and resources}, volume={45}, url={http://dx.doi.org/10.1093/nar/gkw1108}, DOI={10.1093/nar/gkw1108}, number={D1}, journal={Nucleic Acids Research}, publisher={Oxford University Press (OUP)}, year={2016}, month={Nov}, pages={D331–D338}} 
+@article{go2016go, title={Expansion of the Gene Ontology knowledgebase and resources}, volume={45}, url={http://dx.doi.org/10.1093/nar/gkw1108}, DOI={10.1093/nar/gkw1108}, number={D1}, journal={Nucleic Acids Research}, publisher={Oxford University Press (OUP)}, year={2016}, month={Nov}, pages={D331–D338}}
+@article{bostock2011d3,
+  title={D$^3$ data-driven documents},
+  author={Bostock, Michael and Ogievetsky, Vadim and Heer, Jeffrey},
+  journal={IEEE transactions on visualization and computer graphics},
+  volume={17},
+  number={12},
+  pages={2301--2309},
+  year={2011},
+  publisher={IEEE},
+  doi={10.1109/TVCG.2011.185}
+}
+ 
