+++ paper.md
@@ -19,7 +19,7 @@ authors:
   orcid: 0000-0002-7552-1009
   affiliation: 1
 affiliations:
-- name: Bioinformatics and Biostatistics HUB, Centre de Bioinformatique, Biostatistique et Biologie Intégrative (C3BI, USR 3756 Institut Pasteur et CNRS), Paris, France
+- name: Bioinformatics and Biostatistics HUB, Centre de Bioinformatique, Biostatistique et Biologie Intégrative (C3BI, USR 3756 Institut Pasteur CNRS), Paris, France
   index: 1
 - name: French Institute of Bioinformatics, CNRS IFB-Core, Gif-sur-Yvette, France
   index: 2
@@ -29,37 +29,30 @@ bibliography: paper_resources/paper.bib
 
 # Summary
 
-The EDAM Browser is a client-side web-based visualization javascript widget for the navigation of the EDAM ontology [@ison2013edam]. EDAM is a structured and controlled vocabulary describing bioinformatics and computational biology concepts. It is used to annotate resource descriptions (tools, web services) for discovery [@ison2015tools] and integration in platforms such as galaxy [@afgan2016galaxy]. The EDAM Browser is not a generic ontology navigation and/or edition platform, like other systems such as AberOWL [@hoehndorf2015aber], BioPortal [@whetzel2011bioportal], OLS - Ontology Lookup Service [@jupp2015new], Ontobee [@xiang2011ontobee] and WebProt&eacute;g&eacute; [@tudorache2013webprotege]. Rather, it is tailored to the structure and properties of EDAM, and targets both contributors and users who might not be ontology experts. It helps describing resources, and facilitates community contributions to EDAM.
+The EDAM Browser is a client-side web-based visualization javascript widget for the navigation of the EDAM ontology [@ison2013edam]. EDAM is a structured and controlled vocabulary describing bioinformatics and computational biology concepts. It is used to annotate resource descriptions (tools, web services) for discovery [@ison2015tools] and integration in platforms such as galaxy [@afgan2016galaxy]. The EDAM Browser provides users an simple and performant interface to explore EDAM when annotating or searching for bioinformatics resources.
 
-## Availability and re-usablility
+## User interface and functionalities
 
-The EDAM browser is available publicly and anonymously at https://ifb-elixirfr.github.io/edam-browser/. Its lightweight architecture makes it easy to download and run, either as a local HTML file or on a web server. It is possible to integrate the EDAM Browser in external websites and applications, providing a simple way for third party websites to promote EDAM-labeled resources. Both the autocomplete input field and the tree visualization are re-usable: a demonstration code is available [here](https://ifb-elixirfr.github.io/edam-browser/demo.html), showing how the tree can be integrated, how the user can interact with the tree, and how to programmaticaly interact with the tree in JavaScript.
-
-## Information display
-
-The user interface aims at simplicity and relevance to the specific domain of EDAM. The interface displays all of the information necessary to users, and avoids ontology development jargon.
-
-Users can navigate between different parts of the ontology, based on the properties that define their relationships (e.g. this *format* represents this type of *data*, this *data* is an output of this *operation* or is specific of this *topic*). 
[…]
