### 2018-06-15 87d86d3
a few more paper adjustments
#16
