### 2018-06-18 f237d75
Update paper.md
