+++ paper.md
@@ -47,7 +47,7 @@ Additional components of the interface include a full-text search box that let u
 
 ## Architecture, installation and integration
 
-One of the specificities of EDAM is its relatively small size in comparison with large ontologies like Gene Ontology [ashburner2000go, go2016go]. This reduced size makes it easy to load it entirely in the browser's memory, and enables a fast navigation that does not rely on server calls (except for displaying usage statistics from external annotated resources). The architecture of EDAM Browser is therefore very simple: it uses the D3.js library [@bostock2011d3] for visualisation and loads EDAM from a JSON-based format, lighter and easier to parse than the native OWL/XML format. This makes the system easy to download and run, either as a local HTML file or on a web server. It is also possible to integrate it in external websites and applications: the autocomplete input field and the tree visualization are re-usable, and demonstration code is [available online](https://ifb-elixirfr.github.io/edam-browser/demo.html), showing how the tree can be integrated, how the user can interact with the tree, and how to programmaticaly interact with the tree in JavaScript.
+One of the specificities of EDAM is its relatively small size in comparison with large ontologies like Gene Ontology [@ashburner2000go, @go2016go]. This reduced size makes it easy to load it entirely in the browser's memory, and enables a fast navigation that does not rely on server calls (except for displaying usage statistics from external annotated resources). The architecture of EDAM Browser is therefore very simple: it uses the D3.js library [@bostock2011d3] for visualisation and loads EDAM from a JSON-based format, lighter and easier to parse than the native OWL/XML format. This makes the system easy to download and run, either as a local HTML file or on a web server. It is also possible to integrate it in external websites and applications: the autocomplete input field and the tree visualization are re-usable, and demonstration code is [available online](https://ifb-elixirfr.github.io/edam-browser/demo.html), showing how the tree can be integrated, how the user can interact with the tree, and how to programmaticaly interact with the tree in JavaScript.
 
 ## Conclusion
 
