### 2018-06-18 55bb698
Update paper.md
