### 2018-06-18 d658f4f
Update paper.md
