### 2018-06-18 363c1fd
Update paper.md
