+++ paper.md
@@ -35,7 +35,7 @@ The EDAM Browser is a client-side web-based visualization javascript widget for 
 
 The user interface (see Fig. 1) is divided in two main panels, a tree representation of the ontology on the left, and a properties card on the right. The tree on the left represents the different parts of the ontology (formats, data, operations and topics). By expanding or closing the nodes, users explore the hierarchy of the EDAM concepts. The selected concept is displayed in bold, and the hierarchically path leading from the root to it is highlighted. In case this concept has multiple direct or indirect superclasses, all the paths that lead to its different locations in the trees are highlighted (see Fig.1). The properties card panel on the right displays the main properties of the concept, such as its definition, comments, synonyms, and its non-hierachical relations to other concepts (e.g. an operation has inputs, outputs, and belongs to a given topic). It also includes the count and links to entries in various databases that use this term,  such as [bio.tools](https://bio.tools), [BioSphere](https://biosphere.france-bioinformatique.fr/catalogue/) [@brancotte2017biosphere], [BioWeb](http://bioweb.pasteur.fr) and [TeSS](https://tess.elixir-europe.org). 
 
-![The concept has two parents](paper_resources/phylogenetic_tree_generation_with_details.png)
+![EDAM Browser main interface with a concept that have two parents, and relations with others concepts](paper_resources/phylogenetic_tree_generation_with_details.png)
 
 The user may directly suggest either the modification or creation of terms by clicking on the "add a child" and "edit" icons at the top-right corner of this panel. These lead the user to a form that directly formats suggestions as github issues ready to be submitted by the user (see Fig. 2).
 
