### 2018-06-18 c661d1f
Update paper.md
