### 2018-06-18 190b333
Update paper.md
