+++ paper.md
@@ -49,6 +49,6 @@ One of the specificities of EDAM is its relatively small size in comparison with
 
 ## Conclusion
 
-The EDAM Browser is publicly available at https://ifb-elixirfr.github.io/edam-browser/. It aims at simplicity and relevance to EDAM. The interface displays all of the information necessary to users, and avoids ontology development jargon. The EDAM Browser is not a generic ontology navigation and/or edition platform, like other systems such as AberOWL [@hoehndorf2015aber], BioPortal [@whetzel2011bioportal], OLS - Ontology Lookup Service [@jupp2015new], Ontobee [@xiang2011ontobee] and WebProt&eacute;g&eacute; [@tudorache2013webprotege]. Rather, it is tailored to the structure and properties of EDAM, and targets both contributors and users who might not be ontology experts. With this interface, users can both explore quickly and efficiently EDAM, and contribute their own expertise to the ontology. 
+The EDAM Browser is publicly available at https://ifb-elixirfr.github.io/edam-browser/. It aims at providing a simple and efficient entrypoint in the exploration of the EDAM Ontology. The interface displays all of the information necessary to users, and avoids ontology development jargon. The EDAM Browser is not a generic ontology navigation and/or edition platform, like other systems such as AberOWL [@hoehndorf2015aber], BioPortal [@whetzel2011bioportal], OLS - Ontology Lookup Service [@jupp2015new], Ontobee [@xiang2011ontobee] and WebProt&eacute;g&eacute; [@tudorache2013webprotege]. Rather, it is tailored to the structure and properties of EDAM, and targets both contributors and users who might not be ontology experts. With this interface, users can both explore quickly and efficiently EDAM, and contribute their own expertise to the ontology. 
 
 # References
