### 2018-06-18 825399c
Update paper.md
last edits to the paper, based on the suggestions from @alexgarciac in #16
