+++ paper.md
@@ -33,22 +33,22 @@ The EDAM Browser is a client-side web-based visualization javascript widget for 
 
 ## User interface and functionalities
 
-The user interface (see Fig. 1) is divided in two main panels, a tree representation of the ontology on the left, and a properties card on the right. The tree on the left represents the different parts of the ontology (formats, data, operations and topics). By expanding or closing the nodes, users explore the hierarchy of the EDAM concepts. The selected concept is displayed in bold, and the path(s) leading from the root to it is highlighted. In case this concept has multiple direct or indirect superclasses, all the paths that lead to its different locations in the trees are highlighted (see Fig.1). The properties card panel on the right displays the main properties of the concept, such as its definition, comments, synonyms, and its non-hierachical relations to other concepts (e.g. an operation has inputs, outputs, and belongs to a given topic). It also includes counts and links to entries in various databases that use this term, such as [bio.tools](https://bio.tools), [BioSphere](https://biosphere.france-bioinformatique.fr/catalogue/) [@brancotte2017biosphere], [BioWeb](http://bioweb.pasteur.fr) and [TeSS](https://tess.elixir-europe.org). 
+The user interface (see Fig. 1) is divided in two main panels, a tree representation of the ontology on the left, and a properties card on the right. The tree on the left represents the different parts of the ontology (formats, data, operations and topics). By expanding or closing the nodes, users explore the hierarchy of the EDAM concepts. The selected concept is displayed in bold, and all the path(s) leading from the root to it are highlighted. The properties card panel on the right displays the main properties of the concept, such as its definition, comments, synonyms, and its non-hierachical relations to other concepts (e.g. an operation has inputs, outputs, and belongs to a given topic). It also includes counts and links to entries in various databases that use this term, such as [bio.tools](https://bio.tools), [BioSphere](https://biosphere.france-bioinformatique.fr/catalogue/) [@brancotte2017biosphere], [BioWeb](http://bioweb.pasteur.fr) and [TeSS](https://tess.elixir-europe.org). 
 
 ![EDAM Browser main interface with a concept that have two parents, and relations with others concepts](paper_resources/phylogenetic_tree_generation_with_details.png)
 
-The user may directly suggest modification of a term by clicking on the "edit" icons at the top-right corner of this panel ;  creating a new term is also possible thanks to the "add a child" button. Both lead the user to a form that directly formats suggestions as github issues ready to be submitted by the user (see Fig. 2).
+The user may directly suggest a modification on a concept by clicking on the "edit" icons at the top-right corner of this panel ;  creating a new concept is also possible thanks to the "add a child" button. Both lead the user to a form that directly formats suggestions as github issues ready to be submitted by the user (see Fig. 2).
 
 ![The edition form and the issue created](paper_resources/edition_form_to_github_issue.png)
 
-Additional components of the interface include a full-text search box that let users search for concepts on their names, synonyms or descriptions; a navigation toolbar that lets users toggle the display of specific parts of the ontology only, or load locally modified versions of the ontology; and a history stack of the concepts that have been seen previously to the one currently selected, below the properties card panel.
+Additional components of the interface include a full-text search box to search for concepts on their names, synonyms or descriptions; a navigation toolbar to toggle the display of specific parts of the ontology only, or load locally modified versions of the ontology; and a history stack of the concepts that have been seen previously to the one currently selected, below the properties card panel.
 
 ## Architecture, installation and integration
 
[…]
