+++ paper.md
@@ -33,7 +33,7 @@ The EDAM Browser is a client-side web-based visualization javascript widget for 
 
 ## User interface and functionalities
 
-The user interface (see Fig. 1) is divided in two main panels, a tree representation of the ontology on the left, and a properties card on the right. The tree on the left represents the different parts of the ontology (formats, data, operations and topics). By expanding or closing the nodes, users explore the hierarchy of the EDAM concepts. The selected concept is displayed in bold, and all the path(s) leading from the root to it are highlighted. The properties card panel on the right displays the main properties of the concept, such as its definition, comments, synonyms, and its non-hierachical relations to other concepts (e.g. an operation has inputs, outputs, and belongs to a given topic). It also includes counts and links to entries in various databases that use this term, such as [bio.tools](https://bio.tools), [BioSphere](https://biosphere.france-bioinformatique.fr/catalogue/) [@brancotte2017biosphere], [BioWeb](http://bioweb.pasteur.fr) and [TeSS](https://tess.elixir-europe.org). 
+The user interface (see Fig. 1) is divided in two main panels, a tree representation of the ontology on the left, and a properties card on the right. The tree on the left represents the different parts of the ontology (formats, data, operations and topics). By expanding or closing the nodes, users explore the hierarchy of the EDAM concepts. The selected concept is displayed in bold, and all the path(s) leading from the root to it are highlighted. The properties card panel on the right displays the main properties of the concept, such as its definition, comments, synonyms, and its non-hierarchical relations to other concepts (e.g. an operation has inputs, outputs, and belongs to a given topic). It also includes counts and links to entries in various databases that use this term, such as [bio.tools](https://bio.tools), [BioSphere](https://biosphere.france-bioinformatique.fr/catalogue/) [@brancotte2017biosphere], [BioWeb](http://bioweb.pasteur.fr) and [TeSS](https://tess.elixir-europe.org). 
 
 ![EDAM Browser main interface with a concept that have two parents, and relations with others concepts](paper_resources/phylogenetic_tree_generation_with_details.png)
 
