### 2018-07-03 6c7100f
closing #19
