### 2018-07-04 6e33f42
Preparing compatibility with #33
