### 2018-07-04 fb2186a
Fixes #25
