### 2018-07-05 dd844e6
Fixes #28 using levenshtein distance, also splitting query with space and searching for each subpart, and keeping only top 20 results (improve performance)
