### 2018-07-09 22e4b23
improcing pan and zoom
