### 2018-07-18 #698 @alexgarciac
Hi, I have read the latest version and it looks ready to go. @pjotrp answering your question: yes, ready to go into the public domain.
