### 2018-07-18 #698 @alexgarciac
all check boxes (the pertinent ones) on the issue tracker have been checked.
