### 2018-07-18 #698 @pjotrp
Thank you @alexgarciac! :+1: Thank you @bryan-brancotte for your great contribution, there should be more linked data contributions like these.

The review process is now complete. To finalize your submission and accept your paper in JOSS, we need two things. First, can you confirm that
all references in your bibliography have a DOI (if one exists).

Second, we need you to deposit a copy of your software repository (including any revisions made during the JOSS review process) with a data-arc
hiving service. To do so:

 3. Post a comment here with the DOI for the release.
