+++ README.md
@@ -1,6 +1,8 @@
 # Edam stand alone browser using d3.js
 
 [![Build Status](https://travis-ci.org/IFB-ElixirFr/edam-browser.svg?branch=master)](https://travis-ci.org/IFB-ElixirFr/edam-browser)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1314288.svg)](https://doi.org/10.5281/zenodo.1314288)
+
 
 The EDAM Browser is a client-side web-based visualization javascript widget. Its goals are to help describing bio-related resources and service with EDAM, and to facilitate and foster community contributions to EDAM.
 
