### 2018-07-18 #698 @bryan-brancotte
Thanks @alexgarciac @pjotrp for making this contribution being published !
Here is the DOI https://doi.org/10.5281/zenodo.1314288
